etl/requirements.txt
Pillow>=10.2
numpy>=1.26
tqdm>=4.66
